move rest API tests to separate test_* file
